Update Get-DbaCmObject.ps1
I think the UseSSL config value (configured using Set-DbatoolsConfig -FullName 'PSRemoting.PsSession.UseSSL' is not applied when this cmdlet is called.

Background:
Some of our servers are behind a firewall. To get to these we have punched holes for WinRM https port (5986) and configured WinRM on our servers with the relevant thumbprint. For security, we do not allow http (default 5985).

If we try non-dbatools commands like these, it works and we can see the allowed traffic on the firewall:
Test-NetConnection -ComputerName $ComputerName -Port 5986 -Verbose
Test-WSMan -ComputerName $ComputerName -UseSSL -Port 5986 -Verbose
Invoke-Command -ComputerName $ComputerName -UseSSL -Port 5986 -ScriptBlock {Get-WMIObject Win32_ComputerSystem}

If we try the following using dbatools it returns success for CimRM:
Test-DbaCmConnection -ComputerName $ComputerName -Type CimRM -Verbose | Select-Object *

So all good so far... However...
If we try this next command in dbatools, we can see CimRM (WinRM) is attempted before CimDCOM (DCOM) - as expected, but it fails on WinRM. On the firewall we only see blocks for 5985 - as we expect as there is no firewall rule to allow this, but we do not see anything for 5986, which should be the only way to connect seeing as we set UseSSL to true.

So my guess is the underlying code is not factoring the configured dbatools UseSSL value (PSRemoting.PsSession.UseSSL).

Is someone able to take a look at this and fix it please (I'm too fresh-faced with dbatools to suggest a fix)? Can I also suggest the Verbose output specifies if the CimRM connection is using SSL or not too (would make it very clear what's happening here)?